remove wxpython
